Added primitive type loading to C# side

+++ Engine/Source/Scripting/InternalCalls/Resources.cpp
@@ -6,13 +6,13 @@ using namespace std;
 using namespace AquaEngine;
 using namespace AquaEngine::Scripting;
 
-ADD_MANAGED_METHOD(Resource, Duplicate, uint64_t, (unsigned int originalID, MonoString* newPath))
+ADD_MANAGED_METHOD(Resource, Duplicate, uint64_t, (uint64_t originalID, MonoString* newPath))
 {
 	return Resource::IsValidResourceID(originalID) ?
 		Resource::Duplicate(originalID, mono_string_to_utf8(newPath)) : InvalidResourceID;
 }
 
-ADD_MANAGED_METHOD(Resource, GetPath, MonoString*, (unsigned int id))
+ADD_MANAGED_METHOD(Resource, GetPath, MonoString*, (uint64_t id))
 {
 	if (Resource::IsValidResourceID(id))
 		return mono_string_new(mono_domain_get(), Resource::GetPath(id).c_str());
